Moving Console output classes to Elfie.
