+++ Elfie/Elfie.Search/Elfie.Search.csproj
@@ -142,9 +142,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="..\Shared\SharedAssemblyInfo.cs" />
-    <Compile Include="ConsoleHighlighter.cs" />
     <Compile Include="ConsoleInterface.cs" />
-    <Compile Include="Position.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ResultFormatter.cs" />
