+++ Elfie/Elfie/Elfie.csproj
@@ -66,7 +66,9 @@
     <Compile Include="..\Shared\VersionConstants.cs" />
     <Compile Include="..\Shared\SharedAssemblyInfo.cs" />
     <Compile Include="Diagnostics\Cmd.cs" />
+    <Compile Include="Diagnostics\ConsoleHighlighter.cs" />
     <Compile Include="Diagnostics\Logger.cs" />
+    <Compile Include="Diagnostics\Position.cs" />
     <Compile Include="Diagnostics\ProgressWriter.cs" />
     <Compile Include="Extensions\DateTimeExtensions.cs" />
     <Compile Include="Extensions\IntExtensions.cs" />
